feat(metrics): removed old metrics plugin and renamed sales-per-variant
