+++ packages/vendure-plugin-metrics/CHANGELOG.md
@@ -1,37 +1,1 @@
 # Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [4.3.0](https://github.com/Pinelab-studio/pinelab-vendure-plugins/compare/v4.2.3...v4.3.0) (2022-11-11)
-
-### Bug Fixes
-
-- all packages tyo 1.8.3 ([9351d16](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/9351d16d1b4d5bf271dec05d72bd395f86145aae))
-
-### [4.2.2](https://github.com/Pinelab-studio/pinelab-vendure-plugins/compare/v4.2.1...v4.2.2) (2022-10-27)
-
-### Bug Fixes
-
-- **metrics:** export more types ([7e9c6a2](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/7e9c6a2401e21815adb2e686c9b71621b51946ce))
-
-### [4.2.1](https://github.com/Pinelab-studio/pinelab-vendure-plugins/compare/v4.2.0...v4.2.1) (2022-10-27)
-
-### Bug Fixes
-
-- **metrics:** export types ([3216160](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/3216160f68736acb59dd038ba4651fa74965d0a1))
-
-## [4.2.0](https://github.com/Pinelab-studio/pinelab-vendure-plugins/compare/v4.1.2...v4.2.0) (2022-10-27)
-
-### Features
-
-- dynamic chart loading ([5141def](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/5141def41781e39d045920778444f2bfbbc13600))
-- metric plugin chart setup ([fa4c1e5](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/fa4c1e5f03779e94251685443e3a54f92ac848e4))
-- metric plugin chart setup ([584ef9e](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/584ef9eae61af0cfcfd378b7ad9775b606db17f6))
-- metric strategies finished ([98f58e9](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/98f58e93d9d3fcb1813cd96a80b62f1f094d6154))
-- metric strategies finished. still have to implement tests and add readme ([5c0bf25](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/5c0bf2534414f9baaa8aa35ec257f37bf96205d9))
-- metric wip ([db425b4](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/db425b43fab2c09362d54bf631eba006430e33fd))
-- metrics wip API ([8602bb9](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/8602bb9b00044f227069e1b2340d97e604c4229e))
-- **metrics:** finished metrics ([08d62e7](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/08d62e7aa4e2087aa17a5416e713221ce9157c4f))
-- **metrics:** optimalisations ([7e6b773](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/7e6b773de460bf72011c22e927bc910adc022afe))
-- new metric plugin ([8c475cb](https://github.com/Pinelab-studio/pinelab-vendure-plugins/commit/8c475cbf4ef9cabfa126bb673dc09ef66d2c1724))
