+++ packages/vendure-plugin-metrics/package.json
@@ -1,6 +1,6 @@
 {
   "name": "vendure-plugin-metrics",
-  "version": "5.6.7",
+  "version": "5.10.1",
   "description": "Vendure plugin measuring and visualizing e-commerce metrics",
   "author": "Martijn van de Brug <martijn@pinelab.studio>",
   "homepage": "https://pinelab.studio",
@@ -16,7 +16,8 @@
   "scripts": {
     "build": "rimraf dist && yarn graphql-codegen && tsc && copyfiles -u 1 'src/ui/**/*' dist/",
     "test": "jest --preset=\"ts-jest\" --forceExit",
-    "start": "yarn ts-node test/dev-server.ts"
+    "start": "yarn ts-node test/dev-server.ts",
+    "generate": "graphql-codegen"
   },
   "devDependencies": {
     "@graphql-codegen/cli": "^2.4.0",
