+++ packages/vendure-plugin-metrics/src/SalesPerVariantPlugin.ts
