+++ packages/vendure-plugin-metrics/src/api/metrics.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import {
   MetricInterval,
   MetricSummary,
@@ -7,11 +7,10 @@ import {
 } from '../ui/generated/graphql';
 import {
   ConfigService,
-  ListQueryBuilder,
+  ID,
   Logger,
   Order,
   RequestContext,
-  Session,
   TransactionalConnection,
 } from '@vendure/core';
 import {
@@ -22,35 +21,47 @@ import {
   startOfDay,
   sub,
 } from 'date-fns';
-import { loggerCtx, PLUGIN_INIT_OPTIONS } from '../constants';
-import { MetricsPluginOptions } from '../MetricsPlugin';
+import { loggerCtx } from '../constants';
 import { Cache } from './cache';
-import { Between } from 'typeorm';
+import {
+  AverageOrderValueMetric,
+  NrOfTimesSoldMetric,
+  MetricCalculation,
+  NrOfOrdersMetric,
+} from './strategies';
 
 export type MetricData = {
   orders: Order[];
-  sessions: Session[];
 };
 
 @Injectable()
 export class MetricsService {
   cache = new Cache<MetricSummary[]>();
-
+  metricsCalculations: MetricCalculation[];
   constructor(
-    @Inject(PLUGIN_INIT_OPTIONS) private options: MetricsPluginOptions,
-    private listBuilder: ListQueryBuilder,
     private connection: TransactionalConnection,
     private configService: ConfigService
-  ) {}
+  ) {
+    this.metricsCalculations = [
+      new AverageOrderValueMetric(),
+      new NrOfTimesSoldMetric(),
+      new NrOfOrdersMetric(),
+    ];
+  }
 
   async getMetrics(
     ctx: RequestContext,
-    { interval }: MetricSummaryInput
+    { interval, variantIds }: MetricSummaryInput
   ): Promise<MetricSummary[]> {
     // Set 23:59:59.999 as endDate
     const endDate = endOfDay(new Date());
     // Check if we have cached result
-    const cacheKey = { endDate, interval, channel: ctx.channel.token };
+    const cacheKey = {
+      endDate,
+      interval,
+      channel: ctx.channel.token,
+      variantIds,
+    };
     const cachedMetricList = this.cache.get(cacheKey);
     if (cachedMetricList) {
       Logger.info(
@@ -63,12 +74,21 @@ export class MetricsService {
     Logger.info(
       `No cache hit, calculating ${interval} metrics until ${endDate.toISOString()} for channel ${
         ctx.channel.token
+      } for ${
+        variantIds?.length
+          ? `for order containing product variants with ids ${variantIds}`
+          : 'all orders'
       }`,
       loggerCtx
     );
-    const data = await this.loadData(ctx, interval, endDate);
+    const data = await this.loadData(
+      ctx,
+      interval,
+      endDate,
+      variantIds as string[]
+    );
     const metrics: MetricSummary[] = [];
-    this.options.metrics.forEach((metric) => {
+    this.metricsCalculations.forEach((metric) => {
       // Calculate entry (month or week)
       const entries: MetricSummaryEntry[] = [];
       data.forEach((dataPerTick, weekOrMonthNr) => {
@@ -91,10 +111,12 @@ export class MetricsService {
   async loadData(
     ctx: RequestContext,
     interval: MetricInterval,
-    endDate: Date
+    endDate: Date,
+    variantIds?: ID[]
   ): Promise<Map<number, MetricData>> {
     let nrOfEntries: number;
     let backInTimeAmount: Duration;
+    const orderRepo = this.connection.getRepository(ctx, Order);
     // What function to use to get the current Tick of a date (i.e. the week or month number)
     let getTickNrFn: typeof getMonth | typeof getISOWeek;
     let maxTick: number;
@@ -118,52 +140,25 @@ export class MetricsService {
     let hasMoreOrders = true;
     const orders: Order[] = [];
     while (hasMoreOrders) {
-      const [items, nrOfOrders] = await this.listBuilder
-        .build(
-          Order,
-          {
-            skip,
-            take,
-          },
-          {
-            ctx,
-            relations: this.options.orderRelations || [],
-            channelId: ctx.channelId,
-            where: {
-              orderPlacedAt: Between(
-                startDate.toISOString(),
-                endDate.toISOString()
-              ),
-            },
-          }
-        )
-        .getManyAndCount();
+      let query = orderRepo
+        .createQueryBuilder('order')
+        .leftJoinAndSelect('order.lines', 'orderLine')
+        .leftJoin('orderLine.productVariant', 'productVariant')
+        .leftJoinAndSelect('orderLine.items', 'orderItems')
+        .leftJoin('order.channels', 'orderChannel')
+        .where(`orderChannel.id=:channelId`, { channelId: ctx.channelId });
+      if (variantIds?.length) {
+        query = query.andWhere(`productVariant.id IN(:...variantIds)`, {
+          variantIds,
+        });
+      }
+      const [items, nrOfOrders] = await query.getManyAndCount();
       orders.push(...items);
       skip += items.length;
       if (orders.length >= nrOfOrders) {
         hasMoreOrders = false;
       }
     }
-    // Get sessions created or update in given time
-    const sessions: Session[] = [];
-    skip = 0;
-    let hasMoreSessions = true;
-    while (hasMoreSessions) {
-      const [items, nrOfSessions] = await this.connection.rawConnection
-        .getRepository(Session)
-        .findAndCount({
-          select: ['updatedAt'],
-          where: {
-            updatedAt: Between(startDate.toISOString(), endDate.toISOString()),
-          },
-        });
-      sessions.push(...items);
-      skip += items.length;
-      if (sessions.length >= nrOfSessions) {
-        hasMoreSessions = false;
-      }
-    }
-    // Get ticks (weeks or months depending on the interval)
     const dataPerInterval = new Map<number, MetricData>();
     const ticks = [];
     for (let i = 1; i <= nrOfEntries; i++) {
@@ -178,12 +173,8 @@ export class MetricsService {
       const ordersInCurrentTick = orders.filter(
         (order) => getTickNrFn(order.orderPlacedAt!) === tick
       );
-      const sessionsInCurrentTick = sessions.filter(
-        (session) => getTickNrFn(session.updatedAt) === tick
-      );
       dataPerInterval.set(tick, {
         orders: ordersInCurrentTick,
-        sessions: sessionsInCurrentTick,
       });
     });
     return dataPerInterval;
