+++ packages/vendure-plugin-metrics/src/api/schema.graphql.ts
@@ -17,6 +17,7 @@ export const schema = gql`
   }
   input MetricSummaryInput {
     interval: MetricInterval!
+    variantIds: [ID!]
   }
   extend type Query {
     """
