+++ packages/vendure-plugin-metrics/src/constants.ts
@@ -1,3 +1,3 @@
-export const loggerCtx = 'MetricsPlugin';
-export const PLUGIN_INIT_OPTIONS = Symbol('METRICS_PLUGIN_OPTIONS');
-export const PLUGIN_NAME = 'vendure-plugin-metrics';
+export const loggerCtx = 'SalesPerVariantPlugin';
+export const PLUGIN_INIT_OPTIONS = Symbol('SALES_PER_VARIANT_PLUGIN_OPTIONS');
+export const PLUGIN_NAME = 'vendure-plugin-sales-per-variant';
