+++ packages/vendure-plugin-metrics/src/ui/metrics-widget.module.ts
