+++ packages/vendure-plugin-metrics/src/ui/metrics-widget.ts
@@ -1,18 +1,15 @@
-import { Component, NgModule, OnInit } from '@angular/core';
-import { DataService, SharedModule } from '@vendure/admin-ui/core';
+import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
 import {
-  MetricInterval,
-  MetricSummary,
-  MetricSummaryQuery,
-  MetricSummaryQueryVariables,
-} from './generated/graphql';
-import Chart from 'chart.js/auto';
-import { GET_METRICS } from './queries.graphql';
+  DataService,
+  ModalService,
+  ProductMultiSelectorDialogComponent,
+} from '@vendure/admin-ui/core';
+import { MetricInterval, MetricSummary } from './generated/graphql';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { switchMap } from 'rxjs/operators';
+import { MetricsUiService } from './metrics-ui.service';
 
 @Component({
-  selector: 'metrics-widget',
+  selector: 'product-metrics-widget',
   template: `
     <div
       class="btn-group btn-outline-primary btn-sm"
@@ -34,57 +31,111 @@ import { switchMap } from 'rxjs/operators';
       </button>
     </div>
     <br />
-
+    <br />
+    <div>
+      <button
+        (click)="openProductSelectionDialog()"
+        class="btn btn-sm btn-secondary"
+      >
+        {{
+          'common.items-selected-count'
+            | translate : { count: selectedVariantIds?.length ?? 0 }
+        }}...
+      </button>
+      <button
+        class="btn btn-sm circular-button"
+        [attr.disabled]="selectedVariantIds.length == 0 ? 'disabled' : null"
+        (click)="clearProductVariantSelection()"
+      >
+        <clr-icon shape="times"></clr-icon>
+      </button>
+    </div>
     <div *ngFor="let metric of metrics$ | async" class="chart-container">
-      <canvas [id]="metric.code"></canvas>
+      <canvas id="{{ metric.code }}-{{ widgetId }}"></canvas>
     </div>
   `,
   styles: [
     '.chart-container { height: 200px; width: 33%; padding-right: 20px; display: inline-block; padding-top: 20px;}',
     '@media screen and (max-width: 768px) { .chart-container { width: 100%; } }',
+    `
+      .circular-button {
+        width: 16px !important;
+        min-width: 16px !important;
+        height: 16px !important;
+        min-height: 16px !important;
+        border-radius: 50% !important;
+        padding: 0px 0px !important;
+        margin-left: 2.55px !important;
+        margin-bottom: 5px !important;
+        margin-top: 4.25px;
+      }
+    `,
+    `
+      clr-icon[shape='times'] {
+        margin-top: -7.75px !important;
+        margin-left: 0.2px !important;
+      }
+    `,
   ],
 })
 export class MetricsWidgetComponent implements OnInit {
   metrics$: Observable<MetricSummary[]> | undefined;
   charts: any[] = [];
+  variantName: string;
+  dropDownName = 'Select Variant';
+  widgetId = 'product-metrics';
   selection: MetricInterval = MetricInterval.Monthly;
   selection$ = new BehaviorSubject<MetricInterval>(MetricInterval.Monthly);
   nrOfOrdersChart?: any;
-  // Config for all charts
-  config = {
-    x: {
-      grid: {
-        display: false,
-      },
-    },
-    y: {
-      ticks: {
-        display: true,
-      },
-      grid: {
-        display: false,
-        drawBorder: false,
-      },
-    },
-  };
+  selectedVariantIds: string[] = [];
 
-  constructor(private dataService: DataService) {}
+  constructor(
+    private dataService: DataService,
+    private changeDetectorRef: ChangeDetectorRef,
+    private modalService: ModalService,
+    private metricsService: MetricsUiService
+  ) {}
 
   async ngOnInit() {
-    //this.observe();
-    this.metrics$ = this.selection$.pipe(
-      switchMap((selection) => {
-        return this.dataService
-          .query<MetricSummaryQuery, MetricSummaryQueryVariables>(GET_METRICS, {
-            input: {
-              interval: selection,
-            },
-          })
-          .refetchOnChannelChange()
-          .mapStream((metricSummary) => {
-            return metricSummary.metricSummary;
-          });
+    this.loadChartData();
+  }
+
+  onDropdownItemClick(variantId: string, variantName: string) {
+    this.loadChartData();
+    this.dropDownName = variantName;
+  }
+
+  openProductSelectionDialog() {
+    this.modalService
+      .fromComponent(ProductMultiSelectorDialogComponent, {
+        size: 'xl',
+        locals: {
+          mode: 'variant',
+
+          initialSelectionIds: this.selectedVariantIds ?? [],
+        },
       })
+      .subscribe((selection) => {
+        if (selection) {
+          this.selectedVariantIds = [
+            ...selection.map((s) => s.productVariantId),
+          ];
+          this.changeDetectorRef.detectChanges();
+          this.loadChartData();
+        }
+      });
+  }
+
+  clearProductVariantSelection() {
+    this.selectedVariantIds = [];
+    this.changeDetectorRef.detectChanges();
+    this.loadChartData();
+  }
+
+  loadChartData() {
+    this.metrics$ = this.metricsService.queryData(
+      this.selection$,
+      this.selectedVariantIds
     );
     this.metrics$.subscribe(async (metrics) => {
       await new Promise((resolve) => setTimeout(resolve, 100)); // Wait for Angular redraw
@@ -96,35 +147,6 @@ export class MetricsWidgetComponent implements OnInit {
   }
 
   createChart(metric: MetricSummary) {
-    const h = 196; // Vendure hue
-    const s = 100;
-    const l = Math.floor(Math.random() * (80 - 20 + 1)) + 20;
-    const color = h + ', ' + s + '%, ' + l + '%';
-    return new Chart(metric.code, {
-      type: 'bar',
-      data: {
-        // values on X-Axis
-        labels: metric.entries.map((e) => e.label),
-        datasets: [
-          {
-            label: metric.title,
-            data: metric.entries.map((e) => e.value),
-            backgroundColor: `hsla(${color}, 0.4)`,
-            borderColor: `hsla(${color})`,
-            borderWidth: 1,
-          },
-        ],
-      },
-      options: {
-        maintainAspectRatio: false,
-        scales: this.config,
-      },
-    });
+    return this.metricsService.createChart(metric, this.widgetId);
   }
 }
-
-@NgModule({
-  imports: [SharedModule],
-  declarations: [MetricsWidgetComponent],
-})
-export class MetricsWidgetModule {}
