--- packages/vendure-plugin-sales-per-variant/CHANGELOG.md
@@ -1,1 +1,0 @@
-# Change Log
