--- packages/vendure-plugin-sales-per-variant/codegen.yml
@@ -1,9 +1,0 @@
-schema: 'src/**/*.graphql.ts'
-documents: 'src/**/*.graphql.ts'
-generates:
-  ./src/ui/generated/graphql.ts:
-    plugins:
-      - typescript
-      - typescript-operations
-    config:
-      avoidOptionals: false
