--- packages/vendure-plugin-sales-per-variant/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "vendure-plugin-sales-per-variant",
-  "version": "5.10.1",
-  "description": "Vendure plugin measuring and visualizing e-commerce metrics per variant",
-  "author": "Martijn van de Brug <martijn@pinelab.studio>",
-  "homepage": "https://pinelab.studio",
-  "repository": "https://github.com/Pinelab-studio/pinelab-vendure-plugins",
-  "license": "MIT",
-  "private": false,
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "rimraf dist && yarn graphql-codegen && tsc && copyfiles -u 1 'src/ui/**/*' dist/",
-    "test": "jest --preset=\"ts-jest\" --forceExit",
-    "start": "yarn ts-node test/dev-server.ts",
-    "generate": "graphql-codegen"
-  },
-  "devDependencies": {
-    "@graphql-codegen/cli": "^2.4.0",
-    "@graphql-codegen/typescript": "^2.4.8",
-    "@graphql-codegen/typescript-operations": "^2.3.5",
-    "@types/jest": "^27.4.1",
-    "@vendure/admin-ui-plugin": "1.9.5",
-    "@vendure/core": "1.9.5",
-    "@vendure/testing": "1.9.5",
-    "@vendure/ui-devkit": "^1.9.5",
-    "copyfiles": "^2.4.1",
-    "jest": "^27.5.1",
-    "node-fetch": "^2.0.0",
-    "ts-jest": "^27.1.3",
-    "ts-node": "^10.7.0",
-    "typescript": "4.3.5"
-  },
-  "dependencies": {
-    "chart.js": "^3.9.1",
-    "date-fns": "^2.29.3"
-  },
-  "gitHead": "89507cfaa85a9fe1fb204f382f6ba730138fcc27"
-}
