--- packages/vendure-plugin-sales-per-variant/src/api/scalars.graphql.ts
@@ -1,5 +1,0 @@
-import gql from 'graphql-tag';
-// This is only used by codegen so it knows DateTime is a custom scalar
-const scalars = gql`
-  scalar DateTime
-`;
