--- packages/vendure-plugin-sales-per-variant/src/constants.ts
@@ -1,3 +1,0 @@
-export const loggerCtx = 'SalesPerVariantPlugin';
-export const PLUGIN_INIT_OPTIONS = Symbol('SALES_PER_VARIANT_PLUGIN_OPTIONS');
-export const PLUGIN_NAME = 'vendure-plugin-sales-per-variant';
