--- packages/vendure-plugin-sales-per-variant/src/index.ts
@@ -1,5 +1,0 @@
-export * from './SalesPerVariantPlugin';
-export * from './api/strategies';
-export * from './api/metrics.service';
-export * from './api/metrics.resolver';
-export * from './ui/generated/graphql';
