--- packages/vendure-plugin-sales-per-variant/src/ui/metrics-widget.shared-module.ts
@@ -1,21 +1,0 @@
-import { NgModule } from '@angular/core';
-import {
-  registerDashboardWidget,
-  setDashboardWidgetLayout,
-} from '@vendure/admin-ui/core';
-import { MetricsWidgetModule } from './metrics-widget.module';
-
-@NgModule({
-  imports: [MetricsWidgetModule],
-  declarations: [],
-  providers: [
-    registerDashboardWidget('metrics', {
-      title: 'Metrics',
-      supportedWidths: [4, 6, 8, 12],
-      loadComponent: () =>
-        import('./metrics-widget').then((m) => m.MetricsWidgetComponent),
-    }),
-    setDashboardWidgetLayout([{ id: 'metrics', width: 12 }]),
-  ],
-})
-export class MetricsWidgetSharedModule {}
