--- packages/vendure-plugin-sales-per-variant/src/ui/queries.graphql.ts
@@ -1,15 +1,0 @@
-import gql from 'graphql-tag';
-
-export const GET_METRICS = gql`
-  query metricSummary($input: MetricSummaryInput!) {
-    metricSummary(input: $input) {
-      interval
-      code
-      title
-      entries {
-        label
-        value
-      }
-    }
-  }
-`;
