--- packages/vendure-plugin-sales-per-variant/test/dev-server.ts
@@ -1,65 +1,0 @@
-import {
-  createTestEnvironment,
-  registerInitializer,
-  SqljsInitializer,
-} from '@vendure/testing';
-import {
-  DefaultLogger,
-  DefaultSearchPlugin,
-  LogLevel,
-  mergeConfig,
-} from '@vendure/core';
-import { initialData } from '../../test/src/initial-data';
-import { AdminUiPlugin } from '@vendure/admin-ui-plugin';
-import path from 'path';
-import { compileUiExtensions } from '@vendure/ui-devkit/compiler';
-import { SalesPerVariantPlugin } from '../src/';
-import { createSettledOrder } from '../../test/src/shop-utils';
-import { testPaymentMethod } from '../../test/src/test-payment-method';
-
-(async () => {
-  require('dotenv').config();
-  const { testConfig } = require('@vendure/testing');
-  registerInitializer('sqljs', new SqljsInitializer('__data__'));
-  const config = mergeConfig(testConfig, {
-    logger: new DefaultLogger({ level: LogLevel.Debug }),
-    apiOptions: {
-      adminApiPlayground: {},
-      shopApiPlayground: {},
-    },
-    paymentOptions: {
-      paymentMethodHandlers: [testPaymentMethod],
-    },
-    plugins: [
-      SalesPerVariantPlugin,
-      DefaultSearchPlugin,
-      AdminUiPlugin.init({
-        port: 3002,
-        route: 'admin',
-        app: compileUiExtensions({
-          outputPath: path.join(__dirname, '__admin-ui'),
-          extensions: [SalesPerVariantPlugin.ui],
-          devMode: true,
-        }),
-      }),
-    ],
-  });
-  const { server, shopClient } = createTestEnvironment(config);
-  await server.init({
-    initialData: {
-      ...initialData,
-      paymentMethods: [
-        {
-          name: testPaymentMethod.code,
-          handler: { code: testPaymentMethod.code, arguments: [] },
-        },
-      ],
-    },
-    productsCsvPath: '../test/src/products-import.csv',
-  });
-  await createSettledOrder(shopClient, 1);
-  await createSettledOrder(shopClient, 1);
-  await createSettledOrder(shopClient, 1);
-  await createSettledOrder(shopClient, 1);
-  await createSettledOrder(shopClient, 1);
-})();
