--- packages/vendure-plugin-sales-per-variant/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "declaration": true,
-    "outDir": "dist"
-  },
-  "include": ["src/"],
-  "exclude": ["src/ui"]
-}
